Move source to dedicated folder
