+++ Procfile
@@ -1,3 +1,3 @@
 web: ruby -run -e httpd /dev/null -p $PORT
-background: node replication_job.js
+background: node ./src/replication_job.js
 clock: node ./src/schedule-incremental-replication.js
